feat(nvim): disable all default keymaps for SQLComplete
Disable default keybindings for built-in Vim plugin OmniSQL, used by
SQLComplete, due to it also utilized the <C-c> shortcut to access its
functions when editing a SQL file.

Lead to delays when exiting back to normal mode through the
aforementioned key combination, as the plugin would wait for follow up
key presses corresponding to its actions.

To enable a smoother editing experience, this plugin is disabled through
removing its default keybindings, as complete removal is outside scope
due to being a Vim built-in.